Move HttpClient, StorageManager, KeyStore to separate crate
