+++ rust/mobile-toolkit/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "mobile-toolkit"
+version = "0.15.2"
+edition = "2021"
+authors = ["Spruce Systems, Inc."]
+license = "Apache-2.0 OR MIT"
+description = "Shared native interface crate for the Mobile SDK (KeyStore, SigningKey, StorageManagerInterface, and their support types), extracted so UniFFI types unify across separately-generated crates."
+repository = "https://github.com/spruceid/sprucekit-mobile/"
+
+[dependencies]
+anyhow = "1"
+async-trait = "0.1"
+serde = { version = "1.0.204", features = ["derive"] }
+ssi = { version = "0.14", features = ["secp256r1", "secp384r1"] }
+thiserror = "1.0.65"
+uniffi = { version = "0.31", features = ["tokio"] }
