+++ rust/src/crypto.rs
@@ -11,49 +11,11 @@ use isomdl::{
         CoseKey, EC2Curve, EC2Y,
     },
 };
-use serde::{Deserialize, Serialize};
 use ssi::claims::cose::coset;
 
-uniffi::custom_newtype!(KeyAlias, String);
-#[derive(Debug, Serialize, Deserialize, PartialEq, Clone)]
-pub struct KeyAlias(pub String);
-
-#[derive(Debug, uniffi::Error, thiserror::Error)]
-pub enum CryptoError {
-    #[error("{0}")]
-    General(String),
-}
-
-impl CryptoError {
-    fn general(e: impl ToString) -> Self {
-        Self::General(e.to_string())
-    }
-}
-
-impl From<anyhow::Error> for CryptoError {
-    fn from(value: anyhow::Error) -> Self {
-        Self::General(format!("{value:#}"))
-    }
-}
+pub use mobile_toolkit::crypto::{CryptoError, KeyAlias, KeyStore, SigningKey};
 
 type Result<T, E = CryptoError> = ::std::result::Result<T, E>;
-
-#[uniffi::export(with_foreign)]
-/// An interface that can provide access to cryptographic keypairs from the native crypto API.
-pub trait KeyStore: Send + Sync {
-    /// Retrieve a cryptographic keypair by alias. The cryptographic key must be usable for
-    /// creating digital signatures, and must not be usable for encryption.
-    fn get_signing_key(&self, alias: KeyAlias) -> Result<Arc<dyn SigningKey>>;
-}
-
-#[uniffi::export(with_foreign)]
-/// A cryptographic keypair that can be used for signing.
-pub trait SigningKey: Send + Sync {
-    /// Generates a public JWK for this key.
-    fn jwk(&self) -> Result<String>;
-    /// Produces a signature of unknown encoding.
-    fn sign(&self, payload: Vec<u8>) -> Result<Vec<u8>>;
-}
 
 #[derive(uniffi::Object)]
 /// Utility functions for cryptographic curves
@@ -219,8 +181,9 @@ pub(crate) use test::*;
 
 #[cfg(test)]
 mod test {
-    use crate::{local_store::LocalStore, storage_manager::StorageManagerInterface, Key, Value};
+    use crate::local_store::LocalStore;
     use anyhow::Context;
+    use mobile_toolkit::{storage_manager::StorageManagerInterface, Key, Value};
 
     use super::*;
